+++ src/app/router/router.tsx
@@ -3,9 +3,9 @@ import {
   createRoutesFromElements,
   Route,
 } from 'react-router-dom';
-import App from '../App/App';
-import NotFound from '../NotFound/NotFound';
-import ResultDetails from '../Results/ResultDetails/ResultDetails';
+import App from '../App';
+import NotFound from '../../components/NotFound/NotFound';
+import ResultDetails from '../../components/Results/ResultDetails/ResultDetails';
 
 export const router = createBrowserRouter(
   createRoutesFromElements(
