+++ src/components/Results/Results.tsx
@@ -3,7 +3,7 @@ import styles from './Results.module.css';
 import { ReactNode } from 'react';
 import { Outlet } from 'react-router-dom';
 import { Person } from '../../interfaces/SWApi';
-import { LinkWithQuery } from '../Router/LinkWithQuery/LinkWithQuery';
+import { LinkWithQuery } from '../LinkWithQuery/LinkWithQuery';
 import Spinner from '../Spinner/Spinner';
 import NoResults from './NoResults/NoResults';
