+++ src/main.tsx
@@ -2,7 +2,7 @@ import './index.css';
 
 import ReactDOM from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
-import { router } from './components/Router/router';
+import { router } from './app/router/router';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <RouterProvider router={router} />,
